:hammer: chore: fix test
