+++ apps/client/src/components/leo/modals/NameSearchModal/sections/footer-actions.tsx
@@ -2,9 +2,9 @@ import { RecordType } from "@snailycad/types";
 import type { PostEmsFdDeclareCitizenById } from "@snailycad/types/api";
 import { Button } from "@snailycad/ui";
 import { Dropdown } from "components/Dropdown";
-import { normalizeValue } from "context/ValuesContext";
 import { useFeatureEnabled } from "hooks/useFeatureEnabled";
 import useFetch from "lib/useFetch";
+import { normalizeValue } from "lib/values/normalize-value-key";
 import { ThreeDotsVertical } from "react-bootstrap-icons";
 import { useModal } from "state/modalState";
 import { useNameSearch } from "state/search/name-search-state";
