+++ apps/client/src/context/ValuesContext.tsx
@@ -16,6 +16,7 @@ import { ValueType } from "@snailycad/types";
 import type { GetValuesData } from "@snailycad/types/api";
 import { hasValueObj, isBaseValue, isPenalCodeValue } from "@snailycad/utils";
 import type { Router } from "next/router";
+import { normalizeValue } from "lib/values/normalize-value-key";
 
 interface ContextValue<Custom = Value> {
   type: ValueType;
@@ -103,24 +104,4 @@ export function useValues() {
   }
 
   return context;
-}
-
-// transform: PENAL_CODES -> penalCodes
-// transform: DEPARTMENT  -> department
-export function normalizeValue(value: ValueType | (string & {})) {
-  let split = value.toLowerCase().split(/_/);
-
-  if (split.length > 1) {
-    split = split.map((valueType, idx) => {
-      if (idx > 0) {
-        const firstLetter = valueType.charAt(0);
-
-        return [firstLetter.toUpperCase(), valueType.substring(1).toLowerCase()].join("");
-      }
-
-      return valueType.toLowerCase();
-    });
-  }
-
-  return split.join("");
 }
